requirements.txt
# CMSD Extraction Pipeline Dependencies

# PDF Processing
PyMuPDF>=1.23.0
Pillow>=10.0.0

# AI/LLM
anthropic>=0.39.0
openai>=1.0.0

# Database
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
